pkp/pkp-lib#10598 Remove the workflow store duplications across different apps for now as they are common enough
